Make transition across window boundaries smooth.
Used OpenGL translation to make space objects transition smoother when
they move across window boundaries. The whole scene is repeated eight
times in a 3 × 3 matrix around the window. That makes the objects out
of boundaries appear on the opposite side.

This solution will however be unsufficent in the future, when
implementing collision detection. The collision area still won’t cross
the window boundaries.